fix database provisioning and retirement - renames aws databases to aws database - updates specs with correct references to aws database - makes specifying snapshot optional when retiring a database (previous impl. requires the rds to be in the ready state so a snapshot can be taken)
