+++ config/provisioners.json
@@ -1,5 +1,5 @@
 {
-  "AWS Fog Databases": "Jellyfish::Fog::AWS::Databases",
+  "AWS Fog Database": "Jellyfish::Fog::AWS::Database",
   "AWS Fog Storage": "Jellyfish::Fog::AWS::Storage",
   "AWS Fog Infrastructure": "Jellyfish::Fog::AWS::Infrastructure",
   "AWS Fog VPC": "Jellyfish::Fog::AWS::VPC",
