+++ lib/jellyfish_fog.rb
@@ -7,7 +7,7 @@ require 'jellyfish_fog/engine'
 require 'jellyfish_fog/provisioner'
 require 'jellyfish_fog/infrastructure'
 require 'jellyfish_fog/security_group'
-require 'jellyfish_fog/databases'
+require 'jellyfish_fog/database'
 require 'jellyfish_fog/storage'
 require 'jellyfish_fog/subnet'
 require 'jellyfish_fog/vpc'
