+++ spec/jellyfish_fog_spec.rb
@@ -4,7 +4,7 @@ module Jellyfish
   module Fog
     describe 'Registering product types' do
       it 'registers all product types' do
-        expect(Dummy::Application.config.x.product_types).to have_key('AWS Fog Databases')
+        expect(Dummy::Application.config.x.product_types).to have_key('AWS Fog Database')
       end
     end
   end
